refactor: rename controllers and update package structure for clarity
