--- src/main/java/org/thingai/app/scoringservice/ScoringOrchestrator.java
@@ -1,4 +1,0 @@
-package org.thingai.app.scoringservice;
-
-public class ScoringOrchestrator {
-}
