--- src/main/java/org/thingai/app/scoringservice/handler/SyncHandler.java
@@ -1,4 +1,0 @@
-package org.thingai.app.scoringservice.handler;
-
-public class SyncHandler {
-}
