+++ src/main/java/org/thingai/app/scoringservice/ochestrator/ScoringOrchestrator.java
@@ -1,0 +1,4 @@
+package org.thingai.app.scoringservice.ochestrator;
+
+public class ScoringOrchestrator {
+}
